requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
Flask==3.0.3
importlib_metadata==8.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pymongo[srv]==4.8.0
python-dotenv==1.0.1
Werkzeug==3.0.4
zipp==3.20.0
gunicorn==20.1.0
certifi==2024.2.2
